Move 3rd party library to 3rdparty folder
