+++ 3rdparty/minimp3/minimp3.h
