+++ src/mame/machine/k573fpga.h
@@ -12,8 +12,8 @@
 
 #define MINIMP3_ONLY_MP3
 #define MINIMP3_NO_STDIO
-#include "../../devices/sound/minimp3.h"
-#include "../../devices/sound/minimp3_ex.h"
+#include "minimp3/minimp3.h"
+#include "minimp3/minimp3_ex.h"
 
 DECLARE_DEVICE_TYPE(KONAMI_573_DIGITAL_FPGA, k573fpga_device)
